fixing the nesting issue in com.sap.cloud.sdk.cloudplatform.resilience tests
